[fix] (bug): [fixed layout errors]
Need to fix the js navbar file
