+++ src/App.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import Landing from './pages/_Landing';
+import Routes from './routes';
 
 function App() {
   return (
-    <Landing/>
+    <Routes/>
   );
 }
