+++ src/components/BackTriangle/index.js
@@ -8,7 +8,7 @@ import backTriangle from '../../assets/images/Inicio/backTriangle25.svg';
 function BackTriangle() {
   return (
     <div id="box" className="back-box">
-      <img src={backTriangle} className="back-triangle img-fluid" alt="..." />
+      <img src={backTriangle} className="back-triangle img-fluid" alt="#" />
     </div>
   );
 }
