+++ src/pages/Contato/style.css
@@ -77,6 +77,7 @@
   background: var(--color-yellow);
   width: 18rem;
   height: 3.5rem;
+  margin-bottom: 8rem;
 
   display: flex;
   justify-content: center;
@@ -94,7 +95,7 @@
 
 #contato .form .button-box button:hover,
 .modal .button-box button:hover {
-  filter: contrast(1.5);
+  filter: contrast(1);
   width: 25rem;
   box-shadow: 15px 30px 99px rgb(0 0 0 / 16%);
 }
