+++ src/pages/FundosDeInvestimento/style.css
@@ -1,4 +1,7 @@
-@import "../../../pages/Sobre/style.css";
+/* @import "../../../pages/Sobre/style.css"; */
+/* @import "../../pages/"; */
+
+@import "../Sobre/style.css";
 
 .modal {
 	max-width: 100vw;
@@ -15,7 +18,7 @@
 
 .modal .modal-content {
 	background: var(--color-background);
-	background-image: url(../../../assets/images/PortfolioItem/backTriangle75.svg) !important;
+	/* background-image: url(../../../assets/images/PortfolioItem/backTriangle75.svg); */
 	background-repeat: no-repeat;
 	background-position: center center;
 	background-size: 45rem;
@@ -95,7 +98,7 @@
 	height: 100px;
 	background: transparent;
 	filter: contrast(1);
-	background-image: url(../../../assets/images/Modal/fluxo.svg) !important;
+	/* background-image: url(../../../assets/images/Modal/fluxo.svg); */
 	background-repeat: no-repeat;
 	background-position: center center;
 	background-size: 7rem;
@@ -106,12 +109,6 @@
 .modal .image:focus {
 	filter: grayscale(1);
 	background-size: 8rem;
-}
-
-.modal .modal-footer{
-	max-height: 70px;
-	display: flex;
-	align-items: center;
 }
 
 .modal .modal-footer button {
