+++ src/pages/PortfolioItem/index.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 import FundosDeInvestimento from '../../components/Modals/FundosDeInvestimento';
 // import operacoesSecuritizacao from '../../components/Modals/operacoesSecuritizacao';
 // import MergesAcquisitions from '../../components/Modals/MergesAcquisitions';
@@ -16,22 +17,27 @@ function PortfolioItem() {
               <h1 className="display-1">Conheça nosso <br /><span>portfólio</span></h1>
             </div>
             <div className="col-12 col-sm-6 col-lg-3 box">
+
               <button 
-                data-bs-toggle="modal" 
-                data-bs-target="#fundos-de-investimento"
+                // data-bs-toggle="modal" 
+                // data-bs-target="#fundos-de-investimento"
                 // onClick={() => document.getElementById("portfolio-item").scrollIntoView({block: "start", behavior: "instant"})}
                 className="internal-box col-12" 
               >
-                <div className="h1">
-                  Fundos de <br />Investimento
-                </div>
+                <Link to="/fundos-de-investimento" className="give-classes">
+                  <div className="h1">
+                    Fundos de <br />Investimento
+                  </div>
+                </Link>
               </button>
+
             </div>
             <div className="col-12 col-sm-6 col-lg-3 box">
               <button 
-                className="internal-box col-12" 
                 data-bs-toggle="modal" 
-                data-bs-target="#operacoes-securitizacao"
+                data-bs-target="#fundos-de-investimento"
+                // data-bs-target="#operacoes-securitizacao"
+                className="internal-box col-12" 
               >
                 <div className="h1">
                   Operações<br />de Securitização
@@ -40,9 +46,10 @@ function PortfolioItem() {
             </div>
             <div className="col-12 col-sm-6 col-lg-3 box">
               <button 
-                className="internal-box col-12" 
                 data-bs-toggle="modal" 
-                data-bs-target="#merges-and-acquisitions"
+                data-bs-target="#fundos-de-investimento"
+                // data-bs-target="#merges-and-acquisitions"
+                className="internal-box col-12" 
               >
                 <div className="h1">
                   Merges & <br />Acquisitions
@@ -51,9 +58,10 @@ function PortfolioItem() {
             </div>
             <div className="col-12 col-sm-6 col-lg-3 box">
               <button 
-                className="internal-box col-12" 
                 data-bs-toggle="modal" 
-                data-bs-target="#decentralized-finance"
+                data-bs-target="#fundos-de-investimento"
+                // data-bs-target="#decentralized-finance"
+                className="internal-box col-12" 
               >
                 <div className="h1">
                   Decentralized <br />Finance
