+++ src/pages/Sobre/style.css
@@ -182,6 +182,10 @@
   margin-bottom: 0;
 }
 
+#equipe {
+  outline: none;
+}
+
 .accordion-button:not(.collapsed)::after {
   filter: grayscale(1);
 }
